requirements.txt
openai
python-dotenv>=1.0.0
fastapi
uvicorn[standard]>=0.29.0
httpx>=0.24.0
black
pytest
typer
twilio
openai-agents
python-multipart
elevenlabs
slowapi
sqlalchemy
psycopg2-binary>=2.9.0
psycopg[binary,pool]==3.2.11
supabase>=2.4.0
google-api-python-client==2.182.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
hubspot-api-client>=8.0.0
pytz>=2023.3
phonenumbers>=8.13.0
pytest-asyncio>=0.23.0
